+++ src/main/java/com/myproject/video/video_platform/service/product/course/CourseSectionService.java
@@ -8,9 +8,10 @@ import com.myproject.video.video_platform.entity.products.course.CourseSection;
 import com.myproject.video.video_platform.exception.product.ResourceNotFoundException;
 import com.myproject.video.video_platform.repository.products.course.CourseProductRepository;
 import com.myproject.video.video_platform.repository.products.course.CourseSectionRepository;
-import com.myproject.video.video_platform.service.user.UserService;
+import com.myproject.video.video_platform.service.user.CurrentUserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,17 +25,23 @@ public class CourseSectionService {
 
     private final CourseProductRepository courseRepo;
     private final CourseSectionRepository sectionRepo;
-    private final UserService userService;
+    private final CurrentUserService currentUserService;
 
     /**
      * POST /api/products/{productId}/sections
      */
     @Transactional
     public CourseSectionResponseDto createSection(CourseSectionCreateRequestDto dto) {
-
         UUID productId = UUID.fromString(dto.getProductId());
         CourseProduct course = courseRepo.findFullById(productId)
                 .orElseThrow(() -> new ResourceNotFoundException("Course not found: " + productId));
+
+        UUID currentUserId = currentUserService.getCurrentUserId();
+        log.info("User id from context: {}", currentUserId);
+
+
+        if (!course.getUser().getUserId().equals(currentUserId))
+            throw new AccessDeniedException("You don’t own this product.");
 
 
         CourseSection section = new CourseSection();
@@ -66,8 +73,15 @@ public class CourseSectionService {
         CourseSection section = sectionRepo.findById(sectionId)
                 .orElseThrow(() -> new ResourceNotFoundException("Section not found: " + dto.getId()));
 
+        UUID currentUserId = currentUserService.getCurrentUserId();
+        log.info("User id from context: {}", currentUserId);
+
+        if (!section.getCourse().getUser().getUserId().equals(currentUserId))
+            throw new AccessDeniedException("You don’t own this product.");
 
         section.setTitle(dto.getTitle());
+        section.setDescription(dto.getDescription());
+
         if (dto.getPosition() != null) {
             section.setPosition(dto.getPosition());
         }
@@ -77,6 +91,11 @@ public class CourseSectionService {
     public void deleteSection(String id) {
         Optional<CourseSection> sectionOptional = sectionRepo.findById(UUID.fromString(id));
         if (sectionOptional.isPresent()) {
+            UUID currentUserId = currentUserService.getCurrentUserId();
+            log.info("User id from context: {}", currentUserId);
+            if (!sectionOptional.get().getCourse().getUser().getUserId().equals(currentUserId))
+                throw new AccessDeniedException("You don’t own this product.");
+
             sectionRepo.delete(sectionOptional.get());
             log.info("Deleted succesfully a Course section: {}", id);
         } else
