+++ src/main/java/com/myproject/video/video_platform/service/product/strategy_handler/DownloadProductHandler.java
@@ -5,13 +5,16 @@ import com.myproject.video.video_platform.common.enums.products.ProductType;
 import com.myproject.video.video_platform.dto.products.AbstractProductRequestDto;
 import com.myproject.video.video_platform.dto.products.AbstractProductResponseDto;
 import com.myproject.video.video_platform.dto.products.download.DownloadProductRequestDto;
-import com.myproject.video.video_platform.entity.user.User;
 import com.myproject.video.video_platform.entity.products.download.DownloadProduct;
+import com.myproject.video.video_platform.entity.user.User;
 import com.myproject.video.video_platform.exception.product.ResourceNotFoundException;
 import com.myproject.video.video_platform.exception.user.UserNotFoundException;
 import com.myproject.video.video_platform.repository.products.download.DownloadProductRepository;
+import com.myproject.video.video_platform.service.user.CurrentUserService;
 import com.myproject.video.video_platform.service.user.UserService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Component;
 
 import java.util.Optional;
@@ -19,19 +22,14 @@ import java.util.UUID;
 
 @Slf4j
 @Component
+@RequiredArgsConstructor
 public class DownloadProductHandler implements ProductTypeHandler {
 
     private final DownloadProductRepository downloadProductRepository;
     private final DownloadProductConverter downloadProductConverter;
     private final UserService userService;
+    private final CurrentUserService currentUserService;
 
-    public DownloadProductHandler(DownloadProductRepository downloadProductRepository,
-                                  DownloadProductConverter downloadProductConverter,
-                                  UserService userService) {
-        this.downloadProductRepository = downloadProductRepository;
-        this.downloadProductConverter = downloadProductConverter;
-        this.userService = userService;
-    }
 
     @Override
     public ProductType getSupportedType() {
@@ -55,6 +53,12 @@ public class DownloadProductHandler implements ProductTypeHandler {
             throw new UserNotFoundException("User not found with id: " + dto.getUserId());
         }
 
+        UUID currentUserId = currentUserService.getCurrentUserId();
+        log.info("User id from context: {}", currentUserId);
+
+        if (!userOptional.get().getUserId().equals(currentUserId))
+            throw new AccessDeniedException("You don’t own this product.");
+
         DownloadProduct product = downloadProductConverter
                 .mapDownloadProductRequestDtoToEntity((DownloadProductRequestDto) dto, userOptional.get());
         DownloadProduct saved = downloadProductRepository.save(product);
@@ -65,16 +69,19 @@ public class DownloadProductHandler implements ProductTypeHandler {
 
     @Override
     public AbstractProductResponseDto updateProduct(AbstractProductRequestDto dto) {
+        UUID currentUserId = currentUserService.getCurrentUserId();
+        log.info("User id from context: {}", currentUserId);
+
         log.info("Updating a download product: {}", dto.getName());
-        Optional<User> userOptional = userService.findByUserId(UUID.fromString(dto.getUserId()));
-        if (userOptional.isEmpty()) {
-            throw new UserNotFoundException("User not found with id: " + dto.getUserId());
-        }
 
         Optional<DownloadProduct> downloadProductOptional =
                 downloadProductRepository.findById(UUID.fromString(dto.getId()));
+
         if (downloadProductOptional.isPresent()) {
             DownloadProduct downloadProduct = downloadProductOptional.get();
+
+            if (!downloadProduct.getUser().getUserId().equals(currentUserId))
+                throw new AccessDeniedException("You don’t own this product.");
 
             DownloadProduct updatedProduct = downloadProductConverter.mapDownloadProductUpdate(
                     downloadProduct,
@@ -91,8 +98,14 @@ public class DownloadProductHandler implements ProductTypeHandler {
 
     @Override
     public void deleteProduct(String userId, String productId) {
+        UUID currentUserId = currentUserService.getCurrentUserId();
+        log.info("User id from context: {}", currentUserId);
+
         Optional<DownloadProduct> downloadProductOptional = downloadProductRepository.findById(UUID.fromString(productId));
         if (downloadProductOptional.isPresent()) {
+            if (!downloadProductOptional.get().getUser().getUserId().equals(currentUserId))
+                throw new AccessDeniedException("You don’t own this product.");
+
             downloadProductRepository.delete(downloadProductOptional.get());
             log.info("Deleted succesfully a DownloadProduct: {}", productId);
         } else
